+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Display 'Status: Error' on dashboard if a histogram measurement has outliers.
 - Replaced alert() popups with Bootstrap ToastMessages.
 - Added button to clone test configuration in frontend settings.
+- Only export active stream settings on settings export, reducing the size of the exported settings file by up to 90%.
 
 ### Bug fixes
 - Add missing API doc for `POST:/api/ports/` endpoint.